completed solve maze

+++ 1. MicroMouse Code/0. MicroMouse - STM32F407VGT6/PCB_V1/Core/Inc/flood.h
@@ -63,7 +63,7 @@ struct coor
 struct stack
 {
 	struct coor array[256];
-	int index;
+	int16_t index;
 };
 
 // Used to initialize distance array
